moved some methods and variables around
