+++ onboard/godot-frontend/util/Log.cs
@@ -1,21 +1,11 @@
+using System.Dynamic;
 using Godot;
 
 namespace onboard.util; 
 
 public class Log
 {
-    private string className = "null";
-
-    public enum Level
-    {
-        trace,
-        verbose,
-        debug,
-        info,
-        warn,
-        error,
-        fatal,
-    }
+    public string className {get; private set;} = "null";
 
     public Log(string className)
     {
@@ -27,11 +17,11 @@ public class Log
     /// </summary>
     /// <param name="msg"> the message to send </param>
     /// <param name="logLevel"> The level to log </param>
-    public void logMessage(string msg, Level logLevel)
+    public void logMessage(string msg, Logger.Level logLevel)
     {
         string time = Time.GetTimeStringFromSystem();
 
-        string message = $"[{time} {stringWrapLogLevelColor(logLevel.ToString().ToUpperInvariant(), logLevel)} {className}] {msg}";
+        string message = $"[{time} {Logger.stringWrapLogLevelColor(logLevel.ToString().ToUpperInvariant(), logLevel)} {className}] {msg}";
 
         Logger.logMessage(message, logLevel);
     }
@@ -40,61 +30,36 @@ public class Log
     
     public void Trace(string msg)
     {
-        logMessage(msg, Level.trace);
+        logMessage(msg, Logger.Level.trace);
     }
 
     public void Verbose(string msg)
     {
-        logMessage(msg, Level.verbose);
+        logMessage(msg, Logger.Level.verbose);
     }
 
     public void Debug(string msg)
     {
-        logMessage(msg, Level.debug);
+        logMessage(msg, Logger.Level.debug);
     }
 
     public void Info(string msg)
     {
-        logMessage(msg, Level.info);
+        logMessage(msg, Logger.Level.info);
     }
 
     public void Warn(string msg)
     {
-        logMessage(msg, Level.warn);
+        logMessage(msg, Logger.Level.warn);
     }
 
     public void Error(string msg)
     {
-        logMessage(msg, Level.error);
+        logMessage(msg, Logger.Level.error);
     }
 
     public void Fatal(string msg)
     {
-        logMessage(msg, Level.fatal);
-    }
-
-    public static string stringWrapLogLevelColor(string messageToWrap, Level logLevel)
-    {
-        switch(logLevel)
-        {
-            case Level.trace:
-                return $"[color=GREEN]{messageToWrap}[/color]";
-            case Level.verbose:
-                return $"[color=OLD_LACE]{messageToWrap}[/color]";
-            case Level.debug:
-                return $"[color=PURPLE]{messageToWrap}[/color]";
-            case Level.info:
-                return $"[color=GREEN]{messageToWrap}[/color]";
-            case Level.warn:
-                return $"[color=YELLOW]{messageToWrap}[/color]";
-            case Level.error:
-                return $"[color=RED]{messageToWrap}[/color]";
-            case Level.fatal:
-                return $"[color=RED]{messageToWrap}[/color]";
-
-            default:
-                return $"[color=green]{messageToWrap}[/color]";
-                
-        }
+        logMessage(msg, Logger.Level.fatal);
     }
 }
